+++ zsh/dev.zsh
@@ -10,9 +10,13 @@
 alias devcert='generate-ca-and-server --auto ca server 2048'
 
 # PostgreSQL (for use with Nix)
-alias pgst='pg_ctl -l "$PGDATA/server.log" start'
+alias pgst='pg_ctl -o "-k $PGDATA" -l "$PGDATA/server.log" start'
 alias pgsp='pg_ctl stop'
 alias pgswitch='killall postgres && pgst'
+
+# MongoDB (for use with Nix)
+alias mongost='mongod --config mongod.conf --fork'
+alias mongosp='mongo admin --eval "db.shutdownServer()"'
 
 ##
 ## Functions