+++ zsh/django.zsh
@@ -1,0 +1,9 @@
+##################
+# Django Aliases #
+##################
+
+# manage.py
+alias pm='python manage.py'
+alias pmr='python manage.py runserver'
+alias pmt='python manage.py test'
+alias pmsh='python manage.py shell'